requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-adaptors==0.2.5
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
npm==0.1.1
numpy==1.21.2
optional-django==0.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
